Move filter query and paging code to libsplinter
Moves the code used to percent-encode a filter query and provide paging
info for REST API routes to the rest_api module in libsplinter.

The paging code is placed in a new public `splinter::rest_api::paging`
module, and the percent-encoding of the filter query is refactored
behind a single function.
